meetup-infra/custom_resources/requirements.txt
boto3
aws-psycopg2